% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_basis.R
\name{.basis_release_raw_roster}
\alias{.basis_release_raw_roster}
\title{Release Roster Files}
\usage{
.basis_release_raw_roster(seasons)
}
\arguments{
\item{seasons}{The seasons for which to release roster files}
}
\description{
This function releases raw roster files to a release in the nflverse-players
repo. These files serve as base for players data. This function is intended to
update the most recent season in order to get up to date teams and status data.
}
