% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_draft.R
\name{.draft_release_season}
\alias{.draft_release_season}
\title{Release Single Season Draft Data}
\usage{
.draft_release_season(years)
}
\arguments{
\item{years}{One or more draft years to scrape data for. The function will
loop over a vector of years.}
}
\value{
\code{TRUE} invisibly
}
\description{
This function scrapes the PFR draft websites to obtain relevant draft data
by year of the draft. Currently it looks for the following information:
\itemize{
\item{pfr_id (used to join the data to the players dataset)}
\item{draft_year}
\item{draft_round}
\item{draft_pick}
\item{draft_team}
}
}
\details{
The function finishes by releasing a table with the above listed data to the
release tag \code{draft} of the nflverse-players repo.
}
