% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_espn.R
\name{.espn_combine_raw_players}
\alias{.espn_combine_raw_players}
\title{Combine Raw ESPN Player Files}
\usage{
.espn_combine_raw_players(overwrite = !interactive())
}
\arguments{
\item{overwrite}{If \code{TRUE} overwrites all existing raw ESPN files in the
directory \code{./build/raw_espn}.}
}
\value{
Returns the combined raw players ESPN dataset invisibly
}
\description{
This function downloads raw ESPN player files from the release tag \code{raw_espn}
in the nflverse-players repo, and summarizes the data (uses the latest season
of each player) to the raw players ESPN dataset. Raw players data are saved
locally. Set argument \code{overwrite} to \code{TRUE} to always download the latest data.
The function finishes by releasing the raw players ESPN data to the release tag
\code{raw_espn} of the nflverse-players repo.
}
