% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_espn.R
\name{.espn_release_raw_players}
\alias{.espn_release_raw_players}
\title{Release ESPN Players File}
\usage{
.espn_release_raw_players()
}
\description{
This function releases raw ESPN players data to a release in the nflverse-players
repo. It fetches one table with all available athletes from the ESPN API,
compares the returned data with the currently saved data, updates exosting
rows and inserts new rows as needed.
}
