% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_ngs.R
\name{.ngs_release_raw_players}
\alias{.ngs_release_raw_players}
\title{Release NGS Players Files}
\usage{
.ngs_release_raw_players(seasons)
}
\arguments{
\item{seasons}{The seasons for which to release ngs players files}
}
\description{
This function releases raw ngs players files to a release in the nflverse-players
repo. This function is intended to update the most recent season in order
to get up to date ngs player information.
}
