% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_pff.R
\name{.pff_release_raw_players}
\alias{.pff_release_raw_players}
\title{Release PFF Players Files}
\usage{
.pff_release_raw_players(seasons)
}
\arguments{
\item{seasons}{The seasons for which to release pff players files}
}
\description{
This function releases raw pff players files to a release in the nflverse-players
repo. This function is intended to update the most recent season in order
to get up to date pff player information.
}
