% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_basis.R
\name{players_basis_release}
\alias{players_basis_release}
\title{Release Players Basis File}
\usage{
players_basis_release(overwrite = !interactive())
}
\arguments{
\item{overwrite}{If \code{TRUE} overwrites all existing raw roster files in the
directory \code{./build/raw_roster}.}
}
\value{
Returns the players basis dataset invisibly
}
\description{
This function downloads raw roster files from the release tag \code{raw_roster}
in the nflverse-players repo, and summarizes the data (uses the latest season
of each player) to the players basis dataset. Raw rosters are saved locally.
Set argument \code{overwrite} to \code{TRUE} to always download the latest data.
The function finishes by releasing the players basis data to the release tag
\code{players_components} of the nflverse-players repo.
}
