% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_build_dataset.R
\name{players_build_dataset}
\alias{players_build_dataset}
\title{Release Full Players Dataset}
\usage{
players_build_dataset(release = FALSE)
}
\arguments{
\item{release}{If \code{TRUE}, the function will upload the full dataset to the
release tag \code{players_components} of the nflverse-players and nflverse-data
repos.}
}
\value{
The full players dataset invisibly
}
\description{
Download all components of the players dataset and combine them to the full
dataset. The function calls \code{\link[=players_validate]{players_validate()}} internally to make sure, clean
data will be released.
Currently, the following components will be joined:
\itemize{
\item{Basis data, containing basic player info and \code{gsis_id}}
\item{PFR data, adding \code{pfr_id}}
\item{OTC data, adding overthecap player IDs (\code{otc_id}) and \code{pff_id}}
\item{Draft data, scraped from PFR and joined by \code{pfr_id}}
}
}
