% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_download.R
\name{players_download}
\alias{players_download}
\title{Download Players Components}
\usage{
players_download(type = c("basis", "pfr", "otc", "pff", "draft", "full"))
}
\arguments{
\item{type}{One of \code{"basis"}, \code{"pfr"}, \code{"pff"}, \code{"otc"}, \code{"full"}}
}
\value{
A dataframe with players data
}
\description{
Downloads players component files from the release tag \code{players_components}
of the nflverse-players repo
}
