% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_draft.R
\name{players_draft_release}
\alias{players_draft_release}
\title{Release Players Draft File}
\usage{
players_draft_release(overwrite = !interactive())
}
\arguments{
\item{overwrite}{If \code{TRUE} overwrites all existing draft files in the
directory \code{./build/draft}.}
}
\value{
Returns the players draft dataset invisibly
}
\description{
This function downloads yearly draft files from the release tag \code{draft}
in the nflverse-players repo, and summarizes the data (uses the latest season
of each player) to the players draft dataset. Yearly draft data is saved locally.
Set argument \code{overwrite} to \code{TRUE} to always download the latest data.
The function finishes by releasing the players draft data to the release tag
\code{players_components} of the nflverse-players repo.
}
