% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_espn.R
\name{players_espn_release}
\alias{players_espn_release}
\title{Release Players ESPN ID Mapping}
\usage{
players_espn_release(
  players_espn_full_rebuild = Sys.getenv("PLAYERS_ESPN_REBUILD", "false")
)
}
\arguments{
\item{players_espn_full_rebuild}{If \code{FALSE} (the default), the function will
load the players dataset using \code{players_download("full")} and fill only missing
ESPN IDs. Otherwise, it will fill ESPN IDs of all players. Set the environment
variable \code{"PLAYERS_ESPN_REBUILD"} to control the behavior.}
}
\value{
Returns the players ESPN dataset invisibly
}
\description{
This function loads ESPN player IDs and the players basis dataset
(see \link{players_basis_release}) and tries to join as many ESPN IDs as possible.
Currently the function tries to join by
\itemize{
\item{Full name and position}
\item{Full name}
}
Joining is done in the order given and the function cleans the names prior
to joins with \code{\link[nflreadr:clean_player_names]{nflreadr::clean_player_names()}}.
The function finishes by releasing a ESPN ID to GSIS ID mapping to the release
tag \code{players_components} of the nflverse-players repo.
}
\details{
The underlying join function prints messages of join status. You can suppress
the messages by setting \code{options(players_espn_join.verbose = FALSE)}.
}
