% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_manual_overwrite.R
\name{players_fetch_manual_ids}
\alias{players_fetch_manual_ids}
\title{Load a JSON File of Manually Maintained Player IDs}
\usage{
players_fetch_manual_ids()
}
\description{
Parse json file maintained for manual overwrites of player IDs.
}
\details{
The function searches for the file \code{players_manual_overwrite.json} in the
installed packafe directory. If the file doesn't exist for some reason, the
function will download the file from
\url{https://github.com/nflverse/nflverse-players/raw/refs/heads/master/inst/players_manual_overwrite.json}
}
