% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_manual_overwrite.R
\name{players_manual_ids_clean}
\alias{players_manual_ids_clean}
\title{Clean JSON File of Manual IDs}
\usage{
players_manual_ids_clean(manual_ids = players_manual_ids_fetch())
}
\arguments{
\item{manual_ids}{IDs to clean. Defaults to the output of
\code{\link[=players_manual_ids_fetch]{players_manual_ids_fetch()}}}
}
\value{
Cleaned dataset invisibly
}
\description{
Load manually maintained player IDs from JSON file and compare them with
the automatically matched player IDs. If the automated workflow already matched
a player ID that is listed in \code{manual_ids} then that ID will be removed from
the manual ID JSON file.
}
\details{
If the function detects IDs that can be removed, it will ask the user if the
JSON file in \code{inst/players_manual_overwrite.json} should be overwritten.
Changes will be detected from git as the file is version controlled.
}
