% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_manual_overwrite.R
\name{players_manual_overwrite}
\alias{players_manual_overwrite}
\title{Replace Player IDs with Custom IDs}
\usage{
players_manual_overwrite(
  players_df,
  manual_ids = players_fetch_manual_ids(),
  ids_to_replace = overwrite_ids()
)
}
\arguments{
\item{players_df}{A data.frame with the variables \code{gsis_id} and at least the
variables listed in \code{ids_to_replace}}

\item{manual_ids}{A data.frame with the variables \code{gsis_id} and at least the
variables listed in \code{ids_to_replace}. This defaults to the output of
\code{\link[=players_fetch_manual_ids]{players_fetch_manual_ids()}}.}

\item{ids_to_replace}{Variable names where replacement should be performed.
Defaults to gsis_id, pfr_id, pff_id, otc_id, and espn_id}
}
\value{
Returns \code{players_df} with new values from \code{manual_ids}
}
\description{
This function replaces player IDs in \code{players_df} with IDs in \code{manual_ids}.
It replaces the values of the variables \code{ids_to_replace} in \code{players_df} with
their values in \code{manual_ids}. Replacement is only done in the value in
\code{manual_ids} is NOT \code{NA}. Both \code{players_df} and \code{manual_ids} require the
variable \code{gsis_id} for the replacement.
}
