% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_ngs.R
\name{players_ngs_release}
\alias{players_ngs_release}
\title{Release Players NGS File}
\usage{
players_ngs_release(overwrite = !interactive())
}
\arguments{
\item{overwrite}{If \code{TRUE} overwrites all existing raw ngs files in the
directory \code{./build/raw_ngs}.}
}
\value{
Returns the players ngs dataset invisibly
}
\description{
This function downloads raw ngs player files from the release tag \code{raw_ngs}
in the nflverse-players repo, and summarizes the data (uses the latest season
of each player) to the players ngs dataset. Raw players data are saved locally.
Set argument \code{overwrite} to \code{TRUE} to always download the latest data.
The function finishes by releasing the players ngs data to the release tag
\code{players_components} of the nflverse-players repo.
}
