% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_otc.R
\name{players_otc_release}
\alias{players_otc_release}
\title{Release Players Overthecap IDs}
\usage{
players_otc_release(
  players_otc_endpoint = Sys.getenv("PLAYERS_OTC_PLAYERID_ENDPOINT"),
  players_otc_api_key = Sys.getenv("PLAYERS_OTC_API_KEY")
)
}
\arguments{
\item{players_otc_endpoint}{endpoint (defaults to environment variable PLAYERS_OTC_PLAYERID_ENDPOINT)}

\item{players_otc_api_key}{api key (defaults to environment variable PLAYERS_OTC_API_KEY)}
}
\value{
Returns the players OTC dataset invisibly
}
\description{
Thanks to overthecap who kindly grant us access to their player data! This
function queries player IDs from the OTC api and releases them to the release
tag \code{players_components} of the nflverse-players repo.
}
\details{
Please note that the otc players data contain more than only an otc id to gsis
mapping. The data contains the following variables (example values from Eli Manning):
\itemize{
\item{otc_id (865)}
\item{nfl_player_id (2505996)}
\item{esbid (MAN473170)}
\item{gsis_id (00-0022803)}
\item{gsis_it_id (28953)}
\item{pff_id (1722)}
}
}
