% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_pfr.R
\name{players_pfr_release}
\alias{players_pfr_release}
\title{Release Players PFR ID Mapping}
\usage{
players_pfr_release(
  players_pfr_full_rebuild = Sys.getenv("PLAYERS_PFR_REBUILD", "false"),
  players_pfr_source = Sys.getenv("PLAYERS_PFR_BASE")
)
}
\arguments{
\item{players_pfr_full_rebuild}{If \code{FALSE} (the default), the function will
load the players dataset using \code{players_download("full")} and fill only missing
PFR IDs. Otherwise, it will fill PFR IDs of all players. Set the environment
variable \code{"PLAYERS_PFR_REBUILD"} to control the behavior.}

\item{players_pfr_source}{A source URL where we load the PFR players base from.
Set the environment variable \code{"PLAYERS_PFR_BASE"} to control the behavior.}
}
\value{
Returns the players PFR dataset invisibly
}
\description{
This function loads PFR player IDs and the players basis dataset
(see \link{players_basis_release}) and tries to join as many PFR IDs as possible.
Currently the function tries to join by
\itemize{
\item{Name and Rookie Season}
\item{Name}
\item{First, Last}
}
Joining is done in the order given and the function cleans the names prior
to joins with \code{\link[nflreadr:clean_player_names]{nflreadr::clean_player_names()}}.
The function finishes by releasing a PFR ID to GSIS ID mapping to the release
tag \code{players_components} of the nflverse-players repo.
}
\details{
Please note that you have to set the environment variable \code{"PLAYERS_PFR_BASE"}
to the nflverse PFR ID source url. If it is unset, the underlying function
will error.
The underlying join function prints messages of join status. You can suppress
the messages by setting \code{options(players_pfr_join.verbose = FALSE)}.
}
