% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_validate.R
\name{players_validate_extract_replicated}
\alias{players_validate_extract_replicated}
\title{Extract Table of Replicated Data}
\usage{
players_validate_extract_replicated(data, id = relevant_ids())
}
\arguments{
\item{data}{A data frame that was checked with \link{players_validate}}

\item{id}{ID to extract a table of replicates. Must be one of
gsis_id, pfr_id, pff_id, and otc_id}
}
\value{
A table of replicated IDs, sorted by value of \code{id}
}
\description{
Extract Table of Replicated Data
}
