% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_espn.R
\name{.espn_release_raw_players}
\alias{.espn_release_raw_players}
\title{Release ESPN Players Files}
\usage{
.espn_release_raw_players(seasons)
}
\arguments{
\item{seasons}{The seasons for which to release ESPN players files}
}
\description{
This function releases raw ESPN players files to a release in the nflverse-players
repo. This function is intended to update the most recent season in order
to get up to date ESPN player information.
}
