% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_validate.R
\name{players_validate}
\alias{players_validate}
\title{Sanity Checks for Players Data}
\usage{
players_validate(data, verbose = FALSE)
}
\arguments{
\item{data}{A data frame, tibble, or data.table consisting of player IDs.
The variable names must be one of gsis_id, pfr_id, pff_id, and otc_id.}

\item{verbose}{If \code{TRUE}, will print success messages and a summary.}
}
\value{
Returns \code{data}. duplicated observations will be attached in attributes.
}
\description{
Search for duplicated player IDs in a table and attach them to the input data.
The function automatically loops over the player ID variable names
gsis_id, pfr_id, pff_id, and otc_id and searches for duplicated IDs in each
of these variables.
}
\seealso{
\code{\link[=players_validate_extract_duplicated]{players_validate_extract_duplicated()}} to extract tables of
duplicated IDs from attributes.
}
